move files onto folders and create character manager
