+++ src/components/characters/CharTag.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { statusMapping, genderMapping } from "../utils/utils";
+import { statusMapping, genderMapping } from "../../utils/utils";
 
 import "./charTag.css";
