+++ src/components/characters/characters.css
