+++ src/components/header/Header.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import ThemeContext from "../context/ThemeContext";
+import ThemeContext from "../../context/ThemeContext";
 import HeaderBtn from "./HeaderBtn";
 import "./header.css";
